FossilOrigin-Name: 57f16ce8e550f3eac0a22d6c83f24d92885e99cc97229a9550f95984831628fe

+++ src/zip.c
@@ -238,7 +238,7 @@ void zip_set_timedate(double rDate){
   char *zDate = db_text(0, "SELECT datetime(%.17g)", rDate);
   zip_set_timedate_from_str(zDate);
   fossil_free(zDate);
-  unixTime = (rDate - 2440587.5)*86400.0;
+  unixTime = (int)((rDate - 2440587.5)*86400.0);
 }
 
 /*
